python/requirements.txt
google-genai>=1.0.0
fastapi>=0.104.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
uvicorn>=0.24.0
sseclient
